Streamlit code/requirements.txt
streamlit
pytesseract
numpy
torch
Pillow
opencv-python
sentence-transformers
transformers
langchain
langchain-community
langchain-core
langchain-groq
faiss-cpu
unstructured
